Cut down on packet overhead
New breakdown on sizes:

- packet header: 3 bytes (1 src, 1 dst, 1 flags)
- record header: 2 bytes (1 model + 1 record)
- iv: 8 bytes
- data: 21 bytes (4 "0" + 4 "success" + 2 + 0 [] + 2 + 3 "lol" + 2 + 4 "lol2")
- mac: 16 bytes

Total overhead: 29 bytes.
Total: 50 bytes, or x2.38 overhead. That's a lof of overhead.

+++ src/patterns/src/java/app/AppDispatcher.java
@@ -120,8 +120,8 @@ public class AppDispatcher  extends AbstractDispatcher {
                 dst = 11111111;
                 break;
         }
-        pkt.src = src;
-        pkt.dst = dst;
+        pkt.src = (byte)src;
+        pkt.dst = (byte)dst;
         pprint("pkt to send: " + pkt);
 
         return mDriver.sendData(pkt, endpoint);
